[ty] Remove invalid statement-keyword completions in for-statements
In `for x in <CURSOR>` statements it's only valid to provide
expressions that eventually evaluate to an iterable. While
it's extremely difficult to know if something can evaulate to
an iterable in a general case, there are some suggestions we know
can never lead to an iterable. Here we remove keywords
which can never evaluate to an iterable.