Fix `\n` in logging lazy fix in rule G004
Fixes #21973. When the fix is applied and a `\n` is present,
the transformation introduces a newline and results in a syntax error.

The fix replaces `\n` with `\\n` to preserve correct string rendering.

This change also adds support for `Expr::Call` and `Expr::Attribute`
in the fix by replacing the full text range.