doctest fix
Clippy and Cargo want very different things here. Clippy demands that
the numbered list is indented, but that makes Cargo think it's Rust code
and start trying to compile it. I don't know how to make them both
happy, so I'm making it a regular `//` comment block instead of an
actual `///` doc comment, which seems to shut them both up...