Handle all annotations, not just the primary
Also added the severity as a prefix, to emulate what the default
formatter looks like for these messages.

Still haven't quite worked out how to get it to print out a snippet of
the code with arrows/underlines. Maybe leave that for later.

The amount of information in the junit report seems sufficient for a
user to track down the issue, at the moment.