Document more rules with options
Summary
--

While analyzing our rules, I wanted to know which of them use configuration
options but noticed that some of them were not documented (or at least not
documented in a separate `## Options` section).

I had Claude generate an initial list of candidate rules, but it contained a lot
of false positives that I filtered out, and I ended up adding all of these
sections myself. I'm not claiming that the options lists are exhaustive (as in
the rules may use additional options beyond what I found), but this will at
least help with my goal of determining whether or not a rule is configurable at
all and also hopefully be helpful in general.

I mostly just tacked on an `## Options` section without any commentary, but I
added a couple lines of explanation when I felt it wasn't obvious from the
context.

I also noticed a bit of variation in the `flake8-simplify` rules from doing
this. Some of them offer a diagnostic but no fix depending on the resulting line
length of the suggestion, while others offer neither. I'm not sure we need to do
anything different here, but it seemed worth mentioning.

Test Plan
--

Docs tests to make sure the links are right