[ty] Use the AST to suppress keywords in decorators
As of 014abe1ee19e617c13ac498d09de5ae2f932f6e8 the detection
of decorator nodes in the AST is better and we no longer need to
fall back to token-matching to detect decorators. This makes
the logic significantly simpler and fixes some bugs from
inaccuracies in the state machine.